\encoding{utf8}
\name{sscor-package}
\alias{sscor-package}
\docType{package}
\title{
Correlation estimation by spatial signs
}
\description{
R functions for correlation estimation based on spatial signs. Including spatial sign correlation and two stage spatial sign correlation.
}
\details{
\tabular{ll}{
Package: \tab sscor\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-4-22\cr
License: \tab GPL-2 | GPL-3\cr
}
}
\author{
Alexander \enc{Dürre}{Duerre}, Daniel Vogel

Maintainer: Alexander \enc{Dürre}{Duerre} <alexander.duerre@udo.edu>
}
\references{
\enc{Dürre}{Duerre}, A., Vogel, D., Fried, R. (2015): Spatial sign correlation, \emph{Journal of Multivariate Analyis}, vol. 135, 89--105. 
\href{http://arxiv.org/abs/1403.7635}{arvix 1403.7635}

\enc{Dürre}{Duerre}, A., Vogel, D. (2016): Asymptotics of the two-stage spatial sign correlation, \emph{Journal of Multivariate Analyis}, vol. 144, 54--67. \href{http://arxiv.org/abs/1506.02578}{arxiv 1506.02578}

\enc{Dürre}{Duerre}, A., Tyler, D. E., Vogel, D. (2016): On the eigenvalues of the spatial sign covariance matrix in more than two dimensions, to appear in: \emph{Statistics and Probability Letters}. \href{http://arxiv.org/abs/1512.02863}{arvix 1512.02863}
}

